HTTP methods are case sensitive
https://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html#sec5.1.1

According to the HTTP/1.1 specification the method is and must always be
uppercase.  Changed the route data structure to store the methods
uppercase instead of lowercase.

For compatibility reasons we're still normalizing the case of the method
in route() in case anybody was using it instead of the helper methods.